+++ frontend/package.json
@@ -40,6 +40,6 @@
     "eslint-config-next": "15.5.0",
     "prettier": "^3.6.2",
     "tailwindcss": "^4",
-    "typescript": "5.9.2"
+    "typescript": "^5.9.2"
   }
 }
